C++: Skip generating the AddDesc function
We are changing the protobuf compiler and renaming some of the generated
C++ functions. This causes RPCZ to break since RPCZ's generated code
calls the protobuf generated AddDesc function. Here is an example
snippet showing what this looks like:

    void rpcz_protobuf_AddDesc_search_2eproto() {
      static bool already_here = false;
      if (already_here) return;
      already_here = true;
      GOOGLE_PROTOBUF_VERIFY_VERSION;

      ...
    }

However, it appears that the AddDesc function generated by RPCZ is never called
and is not exposed in any header files, so this commit just deletes the RPCZ
code responsible for generating it.